@@ -10,7 +10,7 @@ description: ChameleonMini RevE Android Client.
 version: 0.3.16+29
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,14 +22,11 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
   http: ^0.12.0+2
-  usb_serial:
-    git:
-      url: git://github.com/kgamecarter/usbserial.git
-      ref: master
+  usb_serial: ^0.2.4
   shared_preferences: ^0.5.6
   path_provider: ^1.5.1
-  simple_permissions: ^0.1.9
-  file_picker: ^1.4.3+2
+  permission_handler: ^6.0.1+1
+  file_picker: ^3.0.0
   queries: ^0.1.12
   flutter_nfc_reader:
     git:
